(DI-136) Rename binding dir and package
Renaming the binding dir and package so it can be imported with a go friendly path.

ie. `github.com/puppetlabs/libral/libralgo`
