+++ libralgo/libral.go
@@ -1,4 +1,4 @@
-package libral
+package libralgo
 
 /*
 #cgo LDFLAGS: -fPIC -L/opt/puppetlabs/puppet/lib/ -lral -l:leatherman_json_container.a -l:leatherman_execution.a -l:leatherman_logging.a -l:leatherman_locale.a -l:leatherman_util.a -l:leatherman_file_util.a -laugeas -L /opt/pl-build-tools/lib/ -l:libboost_locale.a -l:libboost_system.a -l:libboost_log.a -l:libboost_log_setup.a -l:libboost_thread.a -l:libboost_date_time.a -l:libboost_filesystem.a -l:libboost_chrono.a -l:libboost_regex.a  -l:libboost_program_options.a -lyaml-cpp -lrt -lxml2 -lz -lfa -lselinux -lstdc++ -lm
@@ -12,8 +12,9 @@ import "C"
 import (
 	"encoding/json"
 	"fmt"
-	"libral/types"
 	"unsafe"
+
+	"github.com/puppetlabs/libral/libralgo/types"
 )
 
 // GetProviders returns the resource providers available to libral.
